+++ src/Network/Ethereum/Web3/Web3.hs
@@ -17,6 +17,7 @@ module Network.Ethereum.Web3.Web3 where
 import           Data.Text                       (Text)
 import           Network.Ethereum.ABI.Prim.Bytes (Bytes)
 import           Network.Ethereum.Web3.Provider  (Web3)
+import           Network.Ethereum.Web3.Types     (Hash)
 import           Network.JsonRpc.TinyClient      (remote)
 
 -- | Returns current node version string.
@@ -25,6 +26,6 @@ clientVersion :: Web3 Text
 clientVersion = remote "web3_clientVersion"
 
 -- | Returns Keccak-256 (not the standardized SHA3-256) of the given data.
-sha3 :: Bytes -> Web3 Bytes
+sha3 :: Bytes -> Web3 Hash
 {-# INLINE sha3 #-}
 sha3 = remote "web3_sha3"