+++ test/Network/Ethereum/Web3/Test/SimpleStorageSpec.hs
@@ -90,7 +90,7 @@ events = describe "can interact with a SimpleStorage contract across block inter
             theSets = [1, 2, 3]
         print "Setting up the filter..."
         fiber <- runWeb3Configured' $ do
-          let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just contractAddress }
+          let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just [contractAddress] }
           forkWeb3 $ processUntil' var fltr ((3 ==) . length)
         print "Setting the values..."
         setValues theCall theSets
@@ -107,9 +107,10 @@ events = describe "can interact with a SimpleStorage contract across block inter
         start <- runWeb3Configured Eth.blockNumber
         let later = BlockWithNumber (start + 3)
             latest = BlockWithNumber (start + 8)
-            fltr = (def :: Filter EvT_CountSet) { filterAddress = Just contractAddress
-                                            , filterFromBlock = later
-                                            , filterToBlock = latest }
+            fltr = (def :: Filter EvT_CountSet) { filterAddress = Just [contractAddress]
+                                                , filterFromBlock = later
+                                                , filterToBlock = latest
+                                                }
         print "Setting up the filter..."
         fiber <- runWeb3Configured' $
           forkWeb3 $ processUntil' var fltr ((3 ==) . length)
@@ -129,18 +130,18 @@ events = describe "can interact with a SimpleStorage contract across block inter
             theSets1 = [7, 8, 9]
             theSets2 = [10, 11, 12]
         start <- runWeb3Configured Eth.blockNumber
-        let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just contractAddress }
+        let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just [contractAddress] }
         fiber <- runWeb3Configured' $ do
           forkWeb3 $ processUntil var fltr ((3 ==) . length) (liftIO . putMVar blockNumberVar . changeBlockNumber)
         print "Running first transactions as past transactions..."
         setValues theCall theSets1
         wait fiber
         print "All past transactions succeeded... "
-        end <- takeMVar blockNumberVar
+        Just end <- takeMVar blockNumberVar
         awaitBlock $ end + 1 -- make past transactions definitively in past
         var' <- newMVar []
         fiber <- runWeb3Configured' $ do
-          let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just contractAddress
+          let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just [contractAddress]
                                               , filterFromBlock = BlockWithNumber start}
           forkWeb3 $ processUntil' var' fltr ((6 ==) . length)
         print "Setting more values"
@@ -157,7 +158,7 @@ events = describe "can interact with a SimpleStorage contract across block inter
             theSets = [4, 5, 6]
         start <- runWeb3Configured Eth.blockNumber
         blockNumberVar <- newEmptyMVar
-        let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just contractAddress }
+        let fltr = (def :: Filter EvT_CountSet) { filterAddress = Just [contractAddress] }
         print "Setting up filter for past transactions..."
         fiber <- runWeb3Configured' $ do
           forkWeb3 $ processUntil var fltr ((3 ==) . length) (liftIO . putMVar blockNumberVar . changeBlockNumber)
@@ -165,7 +166,7 @@ events = describe "can interact with a SimpleStorage contract across block inter
         setValues theCall theSets
         wait fiber
         print "All values have been set"
-        end <- takeMVar blockNumberVar
+        Just end <- takeMVar blockNumberVar
         var' <- newMVar []
         let fltr' = fltr { filterFromBlock = BlockWithNumber start
                          , filterToBlock = BlockWithNumber end