+++ test/Network/Ethereum/Web3/Test/Utils.hs
@@ -35,9 +35,9 @@ import           Data.Traversable                  (for)
 import           GHC.Generics                      (Generic)
 import           Network.Ethereum.ABI.Prim.Address (Address)
 import           Network.Ethereum.Web3.Eth         (accounts, blockNumber)
-import           Network.Ethereum.Web3.Provider    (Provider (..), Web3,
-                                                    Web3Error, runWeb3')
-import           Network.Ethereum.Web3.Types       (BlockNumber, Call (..))
+import           Network.Ethereum.Web3.Provider    (Provider (..), JsonRpcProvider(..)
+                                                   , Web3, Web3Error, runWeb3')
+import           Network.Ethereum.Web3.Types       (Call (..), Quantity)
 import           System.Environment                (lookupEnv, setEnv)
 import           Test.Hspec.Expectations           (shouldSatisfy)
 import           Network.Ethereum.Web3.Net         as Net
@@ -71,14 +71,14 @@ makeContractsEnv = do
 
 runWeb3Configured :: Show a => Web3 a -> IO a
 runWeb3Configured f = do
-    provider <- HttpProvider <$> rpcUri
+    provider <- (flip Provider Nothing . HttpProvider) <$> rpcUri
     v <- runWeb3' provider f
     v `shouldSatisfy` isRight
     let Right a = v in return a
 
 runWeb3Configured' :: Web3 a -> IO a
 runWeb3Configured' f = do
-    provider <- HttpProvider <$> rpcUri
+    provider <- (flip Provider Nothing . HttpProvider) <$> rpcUri
     Right v <- runWeb3' provider f
     return v
 
@@ -101,7 +101,7 @@ sleepSeconds = threadDelay . (* 1000000)
 microtime :: IO Integer
 microtime = numerator . toRational . (* 1000000) <$> getPOSIXTime
 
-awaitBlock :: BlockNumber -> IO ()
+awaitBlock :: Quantity -> IO ()
 awaitBlock bn = do
     bn' <- runWeb3Configured blockNumber
     putStrLn $ "awaiting block " ++ show bn ++ ", currently " ++ show bn'